src/03_Setup-and-Teardown/__test__/03_Scoping.test.ts
import {
  initializeCityDatabase,
  initializeFoodDatabase,
  isCity,
  isValidCityFoodPair,
} from '../03_Scoping';

// Applies to all tests in this file
beforeEach(() => {
  return initializeCityDatabase();
});

test('City database has Vienna', () => {
  expect(isCity('Vienna')).toBeTruthy();
});

test('City database has San Juan', () => {
  expect(isCity('San Juan')).toBeTruthy();
});

describe('Matching cities to foods', () => {
  // Applies only to tests in this describe block
  beforeEach(() => {
    return initializeFoodDatabase();
  });

  test('Vienna <3 veal', () => {
    expect(isValidCityFoodPair('Vienna', 'Wiener Schnitzel')).toBe(true);
  });

  test('San Juan <3 plantains', () => {
    expect(isValidCityFoodPair('San Juan', 'Mofongo')).toBe(true);
  });

  test('Tokyo does not produce Banana', () => {
    expect(isValidCityFoodPair('Tokyo', 'Banana')).toBe(false);
  });
});

// Note that the top-level beforeEach is executed before the beforeEach inside the describe block. It may help to illustrate the order of execution of all hooks.

beforeAll(() => console.log('1 - beforeAll'));
afterAll(() => console.log('1 - afterAll'));
beforeEach(() => console.log('1 - beforeEach'));
afterEach(() => console.log('1 - afterEach'));
test('', () => console.log('1 - test'));
describe('Scoped / Nested block', () => {
  beforeAll(() => console.log('2 - beforeAll'));
  afterAll(() => console.log('2 - afterAll'));
  beforeEach(() => console.log('2 - beforeEach'));
  afterEach(() => console.log('2 - afterEach'));
  test('', () => console.log('2 - test'));
});

// 1 - beforeAll
// 1 - beforeEach
// 1 - test
// 1 - afterEach
// 2 - beforeAll
// 1 - beforeEach
// 2 - beforeEach
// 2 - test
// 2 - afterEach
// 1 - afterEach
// 2 - afterAll
// 1 - afterAll
